Some code refactor of device config types and added config download skeleton
