+++ backend/app/config_download/config_download_service.py
@@ -1,0 +1,2 @@
+class ConfigDownloadService:
+    pass
