+++ backend/app/config_download/dto/download_request_dto.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class ConfigDownloadDto(BaseModel):
+    pass
