+++ backend/app/config_download/utils/file_system_utils.py
@@ -1,0 +1,2 @@
+class FileSystemUtils:
+    pass
