+++ backend/app/running_config/dto/topology_response_dto.py
@@ -1,0 +1,28 @@
+from pydantic import BaseModel
+from enum import Enum
+
+from app.config_upload.util.netmiko_types import NetmikoDeviceType
+
+
+class DeviceType(str, Enum):
+    ROUTER = 'ROUTER'
+    SWITCH = 'SWITCH'
+    UNKNOWN = 'UNKNOWN'
+
+    def to_netmiko_device_type(self):
+        return NetmikoDeviceType.CISCO_IOS if self.value != self.UNKNOWN else NetmikoDeviceType.UNKNOWN
+
+
+class DeviceLink(BaseModel):
+    from_id: str
+    from_if: str
+    to_id: str
+    to_if: str
+
+
+class DeviceConfigInfo(BaseModel):
+    dev_id: str
+    dev_type: DeviceType
+    dev_running_config: list[str]
+    dev_name: str
+    dev_neighbours: list[DeviceLink]
