+++ backend/app/running_config/utils/device_config_constants.py
@@ -1,0 +1,18 @@
+class XmlConfigConstants:
+    PACKET_TRACER_TAG: str = "PACKETTRACER5"
+    NETWORK_TAG: str = "NETWORK"
+    DEVICES_TAG: str = "DEVICES"
+    DEVICE_TAG: str = "DEVICE"
+    ENGINE_TAG: str = "ENGINE"
+    RUNNING_CONFIG_TAG: str = "RUNNINGCONFIG"
+    LINE_TAG: str = "LINE"
+    DEVICE_TYPE_TAG: str = "DEVICE_TYPE"
+    NAME_TAG: str = "NAME"
+    TEXT_TAG: str = "#text"
+    LINKS_TAG: str = "LINKS"
+    LINK_TAG: str = "LINK"
+    CABLE_TAG: str = "CABLE"
+    FROM_TAG: str = "FROM"
+    TO_TAG: str = "TO"
+    PORT_TAG: str = "PORT"
+    SAVE_REF_ID_TAG: str = "SAVE_REF_ID"
